troca dos doubles

+++ objetos_celestes.hpp
@@ -3,22 +3,22 @@
 //||\n\033[0m?
 #include <string>
 
-double calculo_distancia_pela_esfera_de_Hill(double Raio_de_Hill, double excentricidade, double massa_menor, double Massa_maior);
+float calculo_distancia_pela_esfera_de_Hill(float Raio_de_Hill, float excentricidade, float massa_menor, float Massa_maior);
 
 struct Orbita{
-    double apoastro = 0.0;
-    double periastro = 0.0;
-    double excentricidade = 0.0f;
+    float apoastro = 0.0;
+    float periastro = 0.0;
+    float excentricidade = 0.0f;
 
     Orbita();
-    Orbita(const double &apo_ou_peri, const double &excentri, const bool calcular_periastro);
+    Orbita(const float &apo_ou_peri, const float &excentri, const bool calcular_periastro);
 };
 
-Orbita calcula_orbita(double apoastro_ou_periastro, double excentricidade, bool calcular_periastro);
+Orbita calcula_orbita(float apoastro_ou_periastro, float excentricidade, bool calcular_periastro);
 
 class Estrela{
 public:
-    Estrela(std::string &Nome, double &Massa, double &raio, int &Temp, char &c);
+    Estrela(std::string &Nome, float &Massa, float &raio, int &Temp, char &c);
     
     ~Estrela();
 
@@ -27,20 +27,20 @@ public:
 
     std::string get_nome()const;
     Orbita get_orbita()const;
-    double get_massa()const;
-    double get_raio()const;
+    float get_massa()const;
+    float get_raio()const;
     int get_temperatura()const;
     char get_classe()const;
      
 private:
     
-    void define_massa(double m);
-    void define_raio(double r);
+    void define_massa(float m);
+    void define_raio(float r);
     
     std::string m_nome = "";
     Orbita orbt;
-    double m_raio_m = 0.0;
-    double m_massa_Kg = 0.0;
+    float m_raio_m = 0.0;
+    float m_massa_Kg = 0.0;
     int m_temperatura = 0;
     char m_classe = ' ';
 };
@@ -49,24 +49,24 @@ private:
 // em planejamento....
 class Planeta{
 public:
-    Planeta(double m, double r, char c_t, bool roc);
+    Planeta(float m, float r, char c_t, bool roc);
     ~Planeta(){};
     
     std::string nome;
 private:
-    double m_massa_Kg = 0.0;
-    double m_raio_m = 0.0;
+    float m_massa_Kg = 0.0;
+    float m_raio_m = 0.0;
     char m_classe_tamanho = ' ';
     bool m_rochoso = false;
     
-    double m_albedo = 0.0f;
-    double m_atmosfera_densidade = 0.0f;
+    float m_albedo = 0.0f;
+    float m_atmosfera_densidade = 0.0f;
     
 };
 
-Orbita cria_orbitas_aleatorias(std::mt19937 *motor, double menor_valor, double Maior_valor, bool calcula_periastro){
+Orbita cria_orbitas_aleatorias(std::mt19937 *motor, float menor_valor, float Maior_valor, bool calcula_periastro){
     std::uniform_int_distribution<int> excent(10, 60);
-    std::uniform_real_distribution<double> periastro_ou_apoastro(menor_valor, Maior_valor);
+    std::uniform_real_distribution<float> periastro_ou_apoastro(menor_valor, Maior_valor);
     Orbita orbit;
     
     if(calcula_periastro){
@@ -84,10 +84,10 @@ Orbita cria_orbitas_aleatorias(std::mt19937 *motor, double menor_valor, double M
     return orbit;
 }
     
-//double Calculo_Massivo_da_esfera_de_Hill(double distancia, double excentricidade, double massa_menor, double Massa_maior);
-//double Calculo_da_esfera_de_Hill(double distancia, double excentricidade, double massa_menor, double Massa_maior);
+//float Calculo_Massivo_da_esfera_de_Hill(float distancia, float excentricidade, float massa_menor, float Massa_maior);
+//float Calculo_da_esfera_de_Hill(float distancia, float excentricidade, float massa_menor, float Massa_maior);
 
-//double Cauculo_forca_gravitacional(double distancia, double massa1, double massa2);
+//float Cauculo_forca_gravitacional(float distancia, float , float massa2);
 */
 
 #endif
